Restructured code. Moved all unrelated to separate files.
